Adding npm package installation to build
With this change, as we add new development dependencies to this
application, the Docker image will bake these to the environment upon
creation.

To prevent the host filesystem from being polluted with a bunch of
node_modules, the docker-compose definition is updated to overlay a
volume for the node_modules. These will appear only in the container,
not on the host.